docs: 구현 완료된 기능 list 수정
