--- src/main/java/gift/exception/BadRequestExceptionHandler.java
@@ -1,19 +1,0 @@
-package gift.exception;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@RestControllerAdvice
-public abstract class BadRequestExceptionHandler {
-
-    protected ResponseEntity<String> buildResponseEntity(String errorMessage) {
-        return new ResponseEntity<>(errorMessage, HttpStatus.BAD_REQUEST);
-    }
-
-    @ExceptionHandler(Exception.class)
-    public ResponseEntity<String> handleGenericException(Exception ex) {
-        return new ResponseEntity<>("서버 오류 발생.", HttpStatus.INTERNAL_SERVER_ERROR);
-    }
-}
